requirements.txt
a2a-sdk==0.3.4
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contextual-client==0.8.0
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
executing==2.2.1
filetype==1.2.0
fonttools==4.59.2
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.111.0
google-cloud-bigquery==3.36.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.34.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.75
langchain-google-genai==2.1.10
langchain-openai==0.3.32
langchain-redis==0.2.3
langchain-text-splitters==0.3.11
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-checkpoint-redis==0.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.23
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
ml_dtypes==0.5.3
numpy==2.3.2
openai==1.107.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
ply==3.11
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-ulid==3.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
redisvl==0.8.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
zstandard==0.24.0
